+++ src/theme/foundations/radii.ts
@@ -1,4 +1,8 @@
 const radii = {
+  button: {
+    default: "2px",
+    pill: "20px",
+  },
   checkbox: "3px",
   headerDonate: ".25rem",
   pill: "20px",