+++ CHANGELOG.md
@@ -10,9 +10,26 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
+- Adds a variant for `NYCDOE` to the `Logo` component.
+- Adds the `TagSet` component.
+- Adds `AlphabetFilter` component.
+- Adds `AudioPlayer` component.
+
+### Updates
+
+- Updates the `hamburger` SVG for the `Icon` component to allow for color assignments.
+- Updates `ComponentWrapper` prop `descriptionText` type to `string | JSX.Element`.
+
+## 1.1.2 (September 29, 2022)
+
+### Adds
+
 - Adds the `size` prop to the `Button` component. The accepted values are "small",
   "medium", and "large". The default size value is "medium".
-- Adds `AudioPlayer` component.
+- Adds the `value` prop to the `Slider` component to programmatically update the
+  values of the `Slider` component.
+- Adds the `value` prop to the `CheckboxGroup` component to programmatically
+  update the values of the `Checkbox`es within it.
 
 ### Updates
 
@@ -110,7 +127,6 @@ This release is exactly the same as the `1.1.0-rc5` release candidate which was 
 - Updates the `Placeholder` component's CSS to better display in the `Template` components.
 - Updates how HTML attributes are passed down in the `Image` component to the HTML `img` element.
 - Updates the `Accordion` so that its panel content only renders when the `Accordion` is open.
-  > > > > > > > development
 
 ## 1.0.7 (July 29, 2022)